dom/html/test/test_bug1682.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1682
-->
<head>
  <title>Test for Bug 1682</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1682">Mozilla Bug 1682</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 1682 **/
var count = 1;

SimpleTest.waitForExplicitFinish();
addLoadEvent(function () {
  is(count, 1, "onload executes once");
  ++count;
});
addLoadEvent(SimpleTest.finish);




</script>
</pre>
</body>
</html>
